Metaquery DF can support language info as set
Adds a function _make_language_set that converts the string
that is in the 'language' column and converts it into a set.

This is called to create the 'language_set' column.
Having this in the dataframe is duplicate information and in
the future, the function may be left for the user to call.

Not replacing the existing language column to maintain compatibility
with the parent repo.

This allows proper analysis of language distribution.